Wrote test for decimal savings. Fixed HTML form and JS parsing.

+++ imports/ui/save/save.js
@@ -43,7 +43,7 @@ Template.Save.events({
   'submit .add-deposit'(event) {
     event.preventDefault();
     const target = event.target;
-    const amount = parseInt(target.amount.value);
+    const amount = parseFloat(target.amount.value);
     const text = target.text.value;
     if (noAccount()) {
       Meteor.call('savingsAccounts.create');
